requirements.txt
langchain
python-dotenv
langchain-community
openai
prompt-engineering
streamlit
logging